pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autodoc-ai"
version = "0.1.0"
authors = [
  { name="Nirajan Paudel", email="nirajanpaudel33@gmail.com" },
]
description = "An AI-powered tool to automatically document, refactor, and format Python code."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "groq",
    "python-dotenv",
    "pathspec",
    "GitPython",
    "tomli; python_version < '3.11'",
    "black",
]

[project.scripts]
autodoc = "autodoc_ai.cli:main"

# --- Tool's Configuration ---

# This section is read by config.py module at runtime.
[tool.autodoc]
strategy = "groq"
style = "google"
overwrite_existing = false
refactor = false 